+++ airflow/ui/src/components/Clear/index.tsx
